ml/requirements.txt
# Core ML libraries
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
scipy>=1.12.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# Utilities
joblib>=1.4.0
tqdm>=4.66.0
pyyaml>=6.0.0

# MLflow for experiment tracking
mlflow>=2.12.0

# Hyperparameter optimization
optuna>=3.4.0

# Model serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Data validation
great-expectations>=0.18.0

# Additional ML libraries
xgboost>=2.0.0
lightgbm>=4.1.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0